Split json
