+++ frictionless/plugins/json/__init__.py
@@ -1,0 +1,4 @@
+from .dialect import JsonDialect
+from .parser_json import JsonParser
+from .parser_jsonl import JsonlParser
+from .plugin import JsonPlugin
